opspilot-mvp/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
jinja2==3.1.4
google-generativeai==0.8.3
python-dotenv==1.0.1
